+++ src/scripts/compare/dimCompare.directive.js
@@ -31,7 +31,7 @@ function Compare() {
 }
 
 
-function CompareCtrl($scope, toaster, dimCompareService, dimItemService, $translate) {
+function CompareCtrl($scope, toaster, dimCompareService, dimStoreService, $translate) {
   const vm = this;
   vm.tagsEnabled = $featureFlags.tagsEnabled;
   vm.show = dimCompareService.dialogOpen;
@@ -85,7 +85,7 @@ function CompareCtrl($scope, toaster, dimCompareService, dimItemService, $transl
 
     if (args.dupes) {
       vm.compare = args.item;
-      vm.similarTypes = _.where(dimItemService.getItems(), { typeName: vm.compare.typeName });
+      vm.similarTypes = _.where(dimStoreService.getAllItems(), { typeName: vm.compare.typeName });
       let armorSplit;
       if (!vm.compare.location.inWeapons) {
         vm.similarTypes = _.where(vm.similarTypes, { classType: vm.compare.classType });
@@ -106,7 +106,7 @@ function CompareCtrl($scope, toaster, dimCompareService, dimItemService, $transl
           return memo + (stat.base === 0 ? 0 : stat.statHash);
         }, 0) === armorSplit;
       });
-      vm.comparisons = _.where(dimItemService.getItems(), { hash: vm.compare.hash });
+      vm.comparisons = _.where(dimStoreService.getAllItems(), { hash: vm.compare.hash });
     } else if (!_.findWhere(vm.comparisons, { hash: args.item.hash, id: args.item.id })) {
       vm.comparisons.push(args.item);
     }
