+++ src/scripts/services/store/character-utils.js
